+++ src/plugins/unpkg-path-plugin.ts
@@ -1,65 +1,31 @@
 import * as esbuild from "esbuild-wasm";
-import axios from "axios";
-import localForage from "localforage";
 
-const fileCache = localForage.createInstance({
-  name: "fileCache",
-});
 export const unpkgPathPlugin = () => {
   return {
     name: "unpkg-path-plugin",
     setup(build: esbuild.PluginBuild) {
+      // Handle root entry file of 'index.js'
+      build.onResolve({ filter: /(^index\.js$)/ }, () => {
+        return { path: "index.js", namespace: "a" };
+      }); //path to index.js ==> case 1
+
+      // Handle relative paths on a 'module'
+      build.onResolve({ filter: /^\.+\// }, (args: any) => {
+        return {
+          namespace: "a",
+          path: new URL(args.path, "https://unpkg.com" + args.resolveDir + "/")
+            .href,
+        };
+      }); // relative path ==> case 2
+
+      //  Handle main file of a module
       build.onResolve({ filter: /.*/ }, async (args: any) => {
-        console.log("onResolve", args);
-        if (args.path === "index.js") {
-          return { path: args.path, namespace: "a" };
-        }
-
-        if (args.path.includes("./") || args.path.includes("../")) {
-          return {
-            namespace: "a",
-            path: new URL(
-              args.path,
-              "https://unpkg.com" + args.resolveDir + "/"
-            ).href,
-          };
-        }
-
+        // console.log("onResolve", args);
         return {
           namespace: "a",
           path: `https://unpkg.com/${args.path}`,
-        };
+        }; // main file of a module
       }); //onresolve is called whenever esbuild is trying to figure out a path to a particular module
-
-      build.onLoad({ filter: /.*/ }, async (args: any) => {
-        console.log("onLoad", args);
-
-        if (args.path === "index.js") {
-          return {
-            loader: "jsx",
-            contents: `
-              const message = require('react-dom');
-              console.log(message);
-            `,
-          };
-        }
-
-        const cachedResult = await fileCache.getItem<esbuild.OnLoadResult>(
-          args.path
-        );
-        if (cachedResult) {
-          return cachedResult;
-        }
-
-        const { data, request } = await axios.get(args.path);
-        const result: esbuild.OnLoadResult = {
-          loader: "jsx",
-          contents: data,
-          resolveDir: new URL("./", request.responseURL).pathname,
-        };
-        await fileCache.setItem(args.path, result);
-        return result;
-      });
     },
   };
 };
